requirements.txt
bcrypt==3.2.0
boto3==1.16.49
botocore==1.19.49
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
Flask==1.1.1
gevent==20.12.1
greenlet==0.4.17
gunicorn==20.0.4
inotify-simple==1.2.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1607956439537/work
MarkupSafe==1.1.1
mlio @ file:///home/ubuntu/miniconda3/envs/mlio-build/conda-bld/mlio_1601473537821/work/src/mlio-py
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1591485215893/work
paramiko==2.7.2
protobuf==3.14.0
psutil==5.8.0
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
retrying==1.3.3
s3transfer==0.3.3
sagemaker-containers==2.8.6.post2
sagemaker-scikit-learn-extension==2.1.0
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1605985729207/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1604304773994/work
six==1.15.0
threadpoolctl @ file:///tmp/tmp79xdzxkt/threadpoolctl-2.1.0-py3-none-any.whl
typing==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
xgboost==1.3.0
zope.event==4.5.0
zope.interface==5.2.0
